x/escrow/keeper/grpc_query_verdict_test.go
package keeper_test

import (
	"strconv"
	"testing"

	sdk "github.com/cosmos/cosmos-sdk/types"
	"github.com/cosmos/cosmos-sdk/types/query"
	"github.com/stretchr/testify/require"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"

	keepertest "github.com/crow-labs/gamma/testutil/keeper"
	"github.com/crow-labs/gamma/testutil/nullify"
	"github.com/crow-labs/gamma/x/escrow/types"
)

// Prevent strconv unused error
var _ = strconv.IntSize

func TestVerdictQuerySingle(t *testing.T) {
	keeper, ctx := keepertest.EscrowKeeper(t)
	wctx := sdk.WrapSDKContext(ctx)
	msgs := createNVerdict(keeper, ctx, 2)
	for _, tc := range []struct {
		desc     string
		request  *types.QueryGetVerdictRequest
		response *types.QueryGetVerdictResponse
		err      error
	}{
		{
			desc: "First",
			request: &types.QueryGetVerdictRequest{
				CrowId:    msgs[0].CrowId,
				DisputeId: msgs[0].DisputeId,
			},
			response: &types.QueryGetVerdictResponse{Verdict: msgs[0]},
		},
		{
			desc: "Second",
			request: &types.QueryGetVerdictRequest{
				CrowId:    msgs[1].CrowId,
				DisputeId: msgs[1].DisputeId,
			},
			response: &types.QueryGetVerdictResponse{Verdict: msgs[1]},
		},
		{
			desc: "KeyNotFound",
			request: &types.QueryGetVerdictRequest{
				CrowId:    strconv.Itoa(100000),
				DisputeId: strconv.Itoa(100000),
			},
			err: status.Error(codes.NotFound, "not found"),
		},
		{
			desc: "InvalidRequest",
			err:  status.Error(codes.InvalidArgument, "invalid request"),
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			response, err := keeper.Verdict(wctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
				require.Equal(t,
					nullify.Fill(tc.response),
					nullify.Fill(response),
				)
			}
		})
	}
}

func TestVerdictQueryPaginated(t *testing.T) {
	keeper, ctx := keepertest.EscrowKeeper(t)
	wctx := sdk.WrapSDKContext(ctx)
	msgs := createNVerdict(keeper, ctx, 5)

	request := func(next []byte, offset, limit uint64, total bool) *types.QueryAllVerdictRequest {
		return &types.QueryAllVerdictRequest{
			Pagination: &query.PageRequest{
				Key:        next,
				Offset:     offset,
				Limit:      limit,
				CountTotal: total,
			},
		}
	}
	t.Run("ByOffset", func(t *testing.T) {
		step := 2
		for i := 0; i < len(msgs); i += step {
			resp, err := keeper.VerdictAll(wctx, request(nil, uint64(i), uint64(step), false))
			require.NoError(t, err)
			require.LessOrEqual(t, len(resp.Verdict), step)
			require.Subset(t,
				nullify.Fill(msgs),
				nullify.Fill(resp.Verdict),
			)
		}
	})
	t.Run("ByKey", func(t *testing.T) {
		step := 2
		var next []byte
		for i := 0; i < len(msgs); i += step {
			resp, err := keeper.VerdictAll(wctx, request(next, 0, uint64(step), false))
			require.NoError(t, err)
			require.LessOrEqual(t, len(resp.Verdict), step)
			require.Subset(t,
				nullify.Fill(msgs),
				nullify.Fill(resp.Verdict),
			)
			next = resp.Pagination.NextKey
		}
	})
	t.Run("Total", func(t *testing.T) {
		resp, err := keeper.VerdictAll(wctx, request(nil, 0, 0, true))
		require.NoError(t, err)
		require.Equal(t, len(msgs), int(resp.Pagination.Total))
		require.ElementsMatch(t,
			nullify.Fill(msgs),
			nullify.Fill(resp.Verdict),
		)
	})
	t.Run("InvalidRequest", func(t *testing.T) {
		_, err := keeper.VerdictAll(wctx, nil)
		require.ErrorIs(t, err, status.Error(codes.InvalidArgument, "invalid request"))
	})
}
